Corrige ponteiros e animacoes no porte PC

+++ src/battle_anim_mons.c
@@ -335,7 +335,7 @@ void StoreSpriteCallbackInData6(struct Sprite *sprite, void (*callback)(struct S
 
 void SetCallbackToStoredInData6(struct Sprite *sprite)
 {
-    u32 callback = (u16)sprite->data[6] | (sprite->data[7] << 16);
+    u32 callback = (u16)sprite->data[6] | ((u32)(u16)sprite->data[7] << 16);
     sprite->callback = (void (*)(struct Sprite *))callback;
 }
